+++ src/jvmTest/kotlin/choliver/nespot/playtest/CapturingRunner.kt
@@ -1,11 +1,15 @@
-package choliver.nespot.runner
+package choliver.nespot.playtest
 
 import choliver.nespot.cartridge.Rom
 import choliver.nespot.hash
 import choliver.nespot.nes.Joypads
 import choliver.nespot.nes.Nes
+import choliver.nespot.playtest.Scenario.Stimulus
+import choliver.nespot.runner.AudioPlayer
+import choliver.nespot.runner.Event
 import choliver.nespot.runner.Event.*
-import choliver.nespot.runner.Scenario.Stimulus
+import choliver.nespot.runner.KeyAction
+import choliver.nespot.runner.Screen
 import java.util.concurrent.LinkedBlockingQueue
 
 class CapturingRunner(
